Add config knob to suppress Trace log
Probably due to the "beta" status, this mod is *very* chatty.

For example, in one in-game day, this mod emitted 14'000+ mod lines (out
of a total of 15'000 mod lines).

This makes troubleshooting other mods difficult.

The added config.json knob allows player to suppress the Trace log from
this mod.